requirements.txt
requests
Flask
PyGithub
cryptography
gunicorn